https://www.facebook.com/yasmalawi
Because of our interest in schools and education, YAS is reaching out for help in raising $10,700 to support a new YAS project at the Chickhwengwe Primary School in northern Malawi.  In a nutshell, the school is k-8 and serves 280 students in a rural and heavily forested area.  The building, built locally by the community, has no desks and no teacher office and three overworked/underpaid teachers. 
 
The proposal itself is quite simple - funds raised will purchase desks, some building materials, and provide for project oversight/management by YAS.  The bulk of the funds will go to the purchase of desks and building materials to supplement those provided by the community to build an area for the teachers to meet and prepare their lessons.